requirements.txt
PyYAML==6.0.2
watchdog==3.0.0
